scripts/requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
gspread>=5.0.0
pyyaml>=6.0
